requirements.txt
git+git://github.com/Pycord-Development/pycord@master#egg=py-cord
PyYAML>=6.0
mongoengine>=0.23.1
quart>=0.16.1
git+git://github.com/liamdj23/Quart-Discord@master#egg=quart-discord
requests>=2.26.0
qrcode>=7.3.1
Pillow>=8.4.0
pyfiglet>=0.8.post1
beautifulsoup4>=4.10.0
psutil>=5.8.0
PyNaCl>=1.4.0
lavalink>=3.1.5
humanize>=3.13.1